requirements.txt
scrapy==2.13.0
sqlalchemy==2.0.39
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.65.4
chromadb==0.6.3
httpx==0.28.1
